% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.b.r
\name{a.b}
\alias{a.b}
\title{Weight length relationship}
\usage{
a.b(specs, year = ddd, area, season, stat.num = 5, plots = T,
  diet.analysis = F)
}
\arguments{
\item{specs}{what species are included and this is the Maritimes Region Species Codes}

\item{year}{what are the year(s) included in the data}

\item{season}{which season is being analysed? Winter(Dec-Mar), Spring(Apr May), Summer (June to Aug), Autumn (Sept to Dec)}

\item{stat.num}{is the percentile to define the bounds of the length weight regression. From this the number of data points outside the percentile polygon (obtained through bootstrapping) are identified.}

\item{plots}{This will produce your plots of regression and bounding polygon}

\item{diet.analysis}{Uses the info from your diet analysis script to fill in some of the data}
}
\description{
This function uses the data collected through the menu command to estimate as and bs
}
\examples{
diet.analysis()
}

